Feature/update routespages

+++ MDE.Project.Rosseel_Almanzo/MDE.Project.Rosseel_Almanzo/ViewModels/RouteDetailsViewModel.cs
@@ -28,6 +28,27 @@ namespace MDE.Project.Rosseel_Almanzo.ViewModels
         private ObservableCollection<Domain.Models.Image> images;
         private ObservableCollection<Comment> comments;
         private Comment selectedComment;
+        private bool isAdmin;
+        private string commentCreator;
+
+        public string CommentCreator
+        {
+            get => commentCreator;
+            set
+            {
+                commentCreator = value;
+            }
+        }
+
+        public bool IsAdmin
+        {
+            get => isAdmin;
+            set
+            {
+                isAdmin = value;
+                RaisePropertyChanged(nameof(IsAdmin));
+            }
+        }
 
         public Comment SelectedComment
         {
@@ -140,11 +161,15 @@ namespace MDE.Project.Rosseel_Almanzo.ViewModels
             _routesService = routesService;
         }
 
-        public override void Init(object initData)
+        public async override void Init(object initData)
         {
             base.Init(initData);
 
             Id = initData.ToString();
+
+            string admin = await SecureStorage.GetAsync("admin");
+            IsAdmin = bool.Parse(admin);
+            CommentCreator = await SecureStorage.GetAsync("token");
 
             GetRouteDetails.Execute(null);
         }
@@ -240,21 +265,44 @@ namespace MDE.Project.Rosseel_Almanzo.ViewModels
                 return new Command(async () =>
                 {
                     //TODO: Check if comment belongs to logged user!!!!!!!!!!!!!!!!!!!
-
-                    var result = await CoreMethods.DisplayAlert("Delete Comment", "Are u sure u want to delete comment?", "Yes", "Cancel");
-                    if (result)
+                    if(selectedComment.UserId == commentCreator || IsAdmin)
                     {
-                        var deleteResult = await _routesService.DeleteCommentAsync(Id, selectedComment.Id);
-                        if (deleteResult)
+                        var result = await CoreMethods.DisplayAlert("Delete Comment", "Are u sure u want to delete comment?", "Yes", "Cancel");
+                        if (result)
                         {
-                            Comments.Remove(selectedComment);
-                            await CoreMethods.DisplayAlert("Delete Comment", "Comment succesfull deleted", "Ok");
-                        }
-                        else
-                        {
-                            await CoreMethods.DisplayAlert("Delete Comment", "Delete comment failed!", "Ok");
+                            var deleteResult = await _routesService.DeleteCommentAsync(Id, selectedComment.Id);
+                            if (deleteResult)
+                            {
+                                Comments.Remove(selectedComment);
+                                await CoreMethods.DisplayAlert("Delete Comment", "Comment succesfull deleted", "Ok");
+                            }
+                            else
+                            {
+                                await CoreMethods.DisplayAlert("Delete Comment", "Delete comment failed!", "Ok");
+                            }
                         }
                     }
+                });
+            }
+        }
+
+        public ICommand DeleteRouteCommand
+        {
+            get
+            {
+                return new Command(async () =>
+                {
+                    var result = await _routesService.DeleteRouteAsync(Id);
+                    if (result == "Deleted")
+                    {
+                        await CoreMethods.DisplayAlert("Deleted", "Route succesfull deleted!", "Ok");
+                    }
+                    else
+                    {
+                        await CoreMethods.DisplayAlert("Failed", result, "Ok");
+                    }
+
+                    await CoreMethods.PushPageModel<RoutesViewModel>();
                 });
             }
         }
